divide file.
