+++ format/format.go
@@ -4,12 +4,10 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"math"
-	"strconv"
 	"strings"
 	"time"
 
-	"github.com/itchyny/gojq"
+	"github.com/mackerelio/mkr/jq"
 	"github.com/mackerelio/mkr/logger"
 )
 
@@ -33,7 +31,7 @@ func PrettyPrintJSON(outStream io.Writer, src interface{}, query string) error {
 		return err
 	}
 
-	return FilterJSON(outStream, src, query)
+	return jq.FilterJSON(outStream, src, query)
 }
 
 // JSONMarshalIndent call json.MarshalIndent and replace encoded angle brackets
@@ -52,66 +50,4 @@ func replaceAngleBrackets(s string) string {
 func ISO8601Extended(t time.Time) string {
 	const layoutISO8601Exetnded = "2006-01-02T15:04:05-07:00"
 	return t.Format(layoutISO8601Exetnded)
-}
-
-func FilterJSON(outStream io.Writer, src interface{}, queryStr string) error {
-	query, err := gojq.Parse(queryStr)
-	if err != nil {
-		return err
-	}
-
-	var dst interface{}
-	jsonObj, err := json.Marshal(src)
-	if err != nil {
-		return err
-	}
-	err = json.Unmarshal(jsonObj, &dst)
-	if err != nil {
-		return err
-	}
-
-	iter := query.Run(dst)
-	for {
-		v, ok := iter.Next()
-		if !ok {
-			break
-		}
-
-		if err, ok := v.(error); ok {
-			return err
-		}
-
-		if text, e := jsonScalarToString(v); e == nil {
-			_, err := fmt.Fprintln(outStream, text)
-			if err != nil {
-				return err
-			}
-		} else {
-			_, err = fmt.Fprintln(outStream, JSONMarshalIndent(v, "", "    "))
-			if err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// copy from https://github.com/cli/cli/blob/d21d388b8dc10c8f04187c3afa6e0b44f0977c65/pkg/export/template.go#L110-L127
-func jsonScalarToString(input interface{}) (string, error) {
-	switch tt := input.(type) {
-	case string:
-		return tt, nil
-	case float64:
-		if math.Trunc(tt) == tt {
-			return strconv.FormatFloat(tt, 'f', 0, 64), nil
-		} else {
-			return strconv.FormatFloat(tt, 'f', 2, 64), nil
-		}
-	case nil:
-		return "", nil
-	case bool:
-		return fmt.Sprintf("%v", tt), nil
-	default:
-		return "", fmt.Errorf("cannot convert type to string: %v", tt)
-	}
 }
