+++ jq/scalar.go
@@ -1,0 +1,49 @@
+// MIT License
+//
+// Copyright (c) 2019 GitHub Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+package jq
+
+import (
+	"fmt"
+	"math"
+	"strconv"
+)
+
+// copy from https://github.com/cli/cli/blob/d21d388b8dc10c8f04187c3afa6e0b44f0977c65/pkg/export/template.go#L110-L127
+func jsonScalarToString(input interface{}) (string, error) {
+	switch tt := input.(type) {
+	case string:
+		return tt, nil
+	case float64:
+		if math.Trunc(tt) == tt {
+			return strconv.FormatFloat(tt, 'f', 0, 64), nil
+		} else {
+			return strconv.FormatFloat(tt, 'f', 2, 64), nil
+		}
+	case nil:
+		return "", nil
+	case bool:
+		return fmt.Sprintf("%v", tt), nil
+	default:
+		return "", fmt.Errorf("cannot convert type to string: %v", tt)
+	}
+}
